% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDataExists.R
\name{checkDataExists}
\alias{checkDataExists}
\title{Check Data Exists}
\usage{
checkDataExists(id, thisData)
}
\arguments{
\item{id}{Question or Item ID}

\item{thisData}{Dataframe containing data}
}
\value{
TRUE or FALSE
}
\description{
This checks if data exists returning a logical, set to 'dataExists' preferred
}
