% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkipLogic.R
\name{checkSkipLogic}
\alias{checkSkipLogic}
\title{Check Skip Logic}
\usage{
checkSkipLogic(id, blockMap)
}
\arguments{
\item{id}{Question or item id}

\item{blockMap}{blockMap, use blockMap_Item if its for by item description}
}
\value{
TRUE or FALSE
}
\description{
This is a simple skip logic check. If skip logic exists, it displays a message on codebook page, if not, ignores.
Also returns a flag TRUE or FALSE. Assigning to to 'skipLogic'  preferred
}
