% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{cogdata_unnest}
\alias{cogdata_unnest}
\title{Read WearIT Survey data ----}
\usage{
cogdata_unnest(.data)
}
\arguments{
\item{data}{data object returned from \code{cogdata_preprocess()}}
}
\description{
\code{unnest_cogtask_data} unnests M2C2 Cogtask JSON data from WearIT platform
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
