% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R, R/debug.R
\name{cogdata_validation}
\alias{cogdata_validation}
\title{Read WearIT Survey data ----}
\usage{
cogdata_validation(original_data, unnested_data)

cogdata_validation(original_data, unnested_data)
}
\arguments{
\item{original_data}{original, pre-processed data}

\item{unnested_data}{result of \code{unnest_cogtask_data()}}
}
\description{
Read WearIT Survey data ----

Read WearIT Survey data ----
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
