% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexConditionals.R
\name{complexConditional}
\alias{complexConditional}
\title{Complex conditional}
\usage{
complexConditional(
  url =
    "https://wearables.vmhost.psu.edu/wearables-survey/Survey/blockStructure?s_id=75&g_id=1045",
  blockMap
)
}
\arguments{
\item{blockMap}{Blockmap containing data for a single survey}

\item{responseKey}{wearit responseKey}

\item{survey}{Full name of survey you want a flowchart for}

\item{shiny}{boolean indicating whether using shiny app}
}
\value{
updated blockMap
}
\description{
This pulls multi conditional info and joins with blockMap
}
