% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescribeByItems.R
\name{describeByItem}
\alias{describeByItem}
\title{Describe By Item}
\usage{
describeByItem(out2, thisData, blockMap, responseKey)
}
\arguments{
\item{out2}{Character string containing text to be knitted}

\item{thisData}{Dataframe containing data}

\item{blockMap}{df containing blockMap}

\item{responseKey}{df containing responseKey}
}
\value{
out2 or a character string to be knitted
}
\description{
This generates a codebook grouping by items.
That is, if a question is asked across multiple surveys, this will describe the item across all times it appears in a studies dataset
}
