% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescribeByQuestions.R
\name{describeByQuestion}
\alias{describeByQuestion}
\title{Describe By Question}
\usage{
describeByQuestion(out2, thisData, blockMap, responseKey)
}
\arguments{
\item{out2}{Character string containing text to be knitted}

\item{thisData}{Dataframe containing data}

\item{blockMap}{df containing blockMap}

\item{responseKey}{df containing responseKey}
}
\value{
out2 or a character string to be knitted
}
\description{
This generates a codebook by question. That is, a description of every unique instance of a question
If the question appears across multiple surveys, it will display a description of the question for each survey
}
