% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescribeMissingness.R
\name{describeMissingness}
\alias{describeMissingness}
\title{Desribe Missingness}
\usage{
describeMissingness(id, thisData, blockMap)
}
\arguments{
\item{id}{Question or Item ID}

\item{thisData}{Dataframe containing data}

\item{blockMap}{This is needed for the check skip logic call}
}
\value{
cat msg on codebook page
}
\description{
This prints the missingness section for a question or item page
}
