% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkipLogic.R
\name{displaySkipLogic}
\alias{displaySkipLogic}
\title{Display Skip Logic}
\usage{
displaySkipLogic(id, blockMap)
}
\arguments{
\item{id}{Question or item id}

\item{blockMap}{blockMap, use blockMap_Item if its for by item description}
}
\value{
Brief cat of msg on codebook page
}
\description{
This is a wrapper function on checkSkipLogic that displays a msg if its true, does not return the flag
}
