% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequencyTable.R
\name{frequencyTable}
\alias{frequencyTable}
\title{Frequency Table}
\usage{
frequencyTable(id, thisData, responseKey)
}
\arguments{
\item{id}{Question or Item ID}

\item{thisData}{Dataframe containing data}

\item{responseKey}{df containing responseKey}
}
\value{
Frequency table on codebook page
}
\description{
This generates a simple frequency table for multiple select and multiple choice data
}
