% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateBarplot.R
\name{generateBarplot}
\alias{generateBarplot}
\title{Generate Barplot}
\usage{
generateBarplot(id, thisData, questionText)
}
\arguments{
\item{id}{Question or Item id}

\item{thisData}{df containing data}

\item{questionText}{Question or Item's corresponding text squiggly brackets insertion preferred}
}
\value{
ggplot table
}
\description{
Generates a simple bar plot for a given quesiton or item
}
