% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSurveyFlowchart.R
\name{generateSurveyFlowchart}
\alias{generateSurveyFlowchart}
\title{Generate Survey Flowchart}
\usage{
generateSurveyFlowchart(blockMap, responseKey, survey, shiny = FALSE)
}
\arguments{
\item{blockMap}{Blockmap containing data for a single survey}

\item{responseKey}{wearit responseKey}

\item{survey}{Full name of survey you want a flowchart for}

\item{shiny}{boolean indicating whether using shiny app}
}
\value{
Mermaid flowchart
}
\description{
This generates a Survey Flowchart for a studies WearIT data
}
