% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshots.R
\name{getSurveyScreenshots}
\alias{getSurveyScreenshots}
\title{screenshots}
\usage{
getSurveyScreenshots(
  survey,
  blockMap = read.csv("blockMap.csv"),
  responseKey = read.csv("responseKey.csv"),
  shiny = FALSE
)
}
\arguments{
\item{survey}{Full name of survey you want a flowchart for}

\item{blockMap}{Blockmap containing data for a single survey}

\item{responseKey}{wearit responseKey}

\item{shiny}{boolean indicating whether using shiny app}
}
\description{
This generates yaml file for maestro to grab wear it screenshots
}
