% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListQuestionIds.R
\name{listQuestionIds}
\alias{listQuestionIds}
\title{List Question Ids}
\usage{
listQuestionIds(id, blockMap)
}
\arguments{
\item{id}{Item ID number}

\item{blockMap}{df containing blockMap}
}
\value{
Text on the item pages
}
\description{
This lists the question ids associated with an item as a tabset so that you can link back out to question page
Put results = 'asis' and echo = FALSE in codechunk block
}
