% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHelpers.R
\name{loadSavedData}
\alias{loadSavedData}
\title{loadSavedData}
\usage{
loadSavedData()
}
\arguments{
\item{shiny}{Boolean indicating whether the user is in the shiny app or not}
}
\value{
blockMap Data and responseKey in environment
}
\description{
This loads all WearIT data
}
