% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentChild.R
\name{parentChild}
\alias{parentChild}
\title{parentChild}
\usage{
parentChild(blockMap)
}
\arguments{
\item{blockMap}{dataframe containing blockMap}
}
\value{
TRUE or FALSE cols in blockMap
}
\description{
This makes 2 new boolean columns in blockMap, parent and child.
If the block or question is a parent, TRUE it, otherwise FALSE. Same for child. This does not check for parent or child of survey.
}
