% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHelpers.R
\name{parseBySurvey}
\alias{parseBySurvey}
\title{Parse by Survey}
\usage{
parseBySurvey(survey, all = FALSE)
}
\arguments{
\item{survey}{Name of the survey to pull ex. "Daily Diary"}

\item{all}{Whether to pull a different blockMap for each survey... not implemented yet}

\item{shiny}{Boolean indicating whether using shiny app}
}
\value{
A block map for a given "survey"
}
\description{
This helps grab a blockMap for a single survey.
}
