% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SliderDescriptives.R
\name{sliderDescriptives}
\alias{sliderDescriptives}
\title{Slider Descriptives}
\usage{
sliderDescriptives(id, thisData)
}
\arguments{
\item{id}{Question or Item ID}

\item{thisData}{Dataframe containing data}
}
\value{
cat msg on codebook page
}
\description{
This prints descriptives for a slider item or question type on the codebook page
}
