% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SliderHist.R, R/SliderScale.R
\name{sliderHist}
\alias{sliderHist}
\title{Silder Hist}
\usage{
sliderHist(id, thisData)

sliderHist(id, thisData)
}
\arguments{
\item{id}{Question or Item ID}

\item{thisData}{Dataframe containing data}
}
\value{
generates basic hist for continuous data

generates basic hist for continuous data
}
\description{
This function generates a basic histogram for a slider Item or Question

This function generates a basic histogram for a slider Item or Question
}
