% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCodebook.R
\name{generateCodebook}
\alias{generateCodebook}
\title{Generate Codebook}
\usage{
generateCodebook(
  title = "This is a temp default title, please set",
  authors = "This is a temp default author list, please set",
  funding = "This is a temp default funding source, please set",
  abstract = "This is a temp default abstract, please set",
  summary = "This is a temp default summary, please set",
  thisData = read.csv("Codebook_RMD/Data/Data.csv"),
  blockMap = read.csv("Codebook_RMD/Data/blockMap.csv"),
  responseKey = read.csv("Codebook_RMD/Data/responseKey.csv"),
  include_practice_items = FALSE,
  include_html = FALSE,
  print_blockmap_missing = FALSE,
  generate_item_description = FALSE,
  shiny = FALSE,
  codebookChunkDisplayOptions = "",
  output_file = "Codebook.html"
)
}
\arguments{
\item{title}{Character string denoting the desired title}

\item{authors}{Character string listing study or codebook authors}

\item{funding}{Character string denoting funding sources}

\item{abstract}{Character string containing abstract of study}

\item{summary}{Character string containing summary of study findings}

\item{thisData}{Dataframe containing data}

\item{blockMap}{Dataframe containing blockMap}

\item{responseKey}{Dataframe containing responseKey}

\item{include_practice_items}{Logical indicating whether to include practice item in the data description - note... this is not supported currently}

\item{include_html}{Logical indicating whether to keep html formatting - note... html formatting for question text not supported currently}

\item{print_blockmap_missing}{Logical. If true, will display the rows with NA values for Question.Type.Display.Name of which, will be automatically removed}

\item{generate_item_description}{Logical indicating whether to include by item description}
}
\value{
codebook.html file in Codebook_RMD folder
}
\description{
This generates a codebook for a studies WearIT data
}
