% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDataDictionary.R
\name{generateDataDictionary}
\alias{generateDataDictionary}
\title{Generate Data Dictionary}
\usage{
generateDataDictionary(
  blockMap = read.csv(get_resource_path("Codebook_RMD", "Data", "blockMap.csv")),
  responseKey = read.csv(get_resource_path("Codebook_RMD", "Data", "responseKey.csv")),
  ...
)
}
\value{
datadictionary.csv file in Codebook_RMD folder
}
\description{
This generates a data dictionary for a studies WearIT data
}
