% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyTree.R
\name{generateSurveyTree}
\alias{generateSurveyTree}
\title{Generate Survey Tree}
\usage{
generateSurveyTree(shiny = FALSE, blockMapParsed)
}
\arguments{
\item{shiny}{boolean indicating whether using shiny app}

\item{blockMap}{Blockmap containing data for a single survey}
}
\value{
survey tree
}
\description{
This generates a Survey Tree for a studies WearIT data
}
