% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyTree backup.R, R/SurveyTree.R
\name{generateSurveyTree}
\alias{generateSurveyTree}
\title{Generate Survey Tree}
\usage{
generateSurveyTree(shiny = FALSE, blockMapParsed, showLabels = TRUE)

generateSurveyTree(shiny = FALSE, blockMapParsed, showLabels = TRUE)
}
\arguments{
\item{shiny}{boolean indicating whether using shiny app}

\item{blockMapParsed}{Blockmap containing data for a single survey}

\item{showLabels}{boolean indicating whether to show node labels}

\item{blockMap}{Blockmap containing data for a single survey}
}
\value{
survey tree

survey tree rendered as a Mermaid flowchart
}
\description{
This generates a Survey Tree for a studies WearIT data

This generates a Survey Tree for a study's WearIT data using DiagrammeR/Mermaid
}
