% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyTree.R
\name{generateSurveyTree}
\alias{generateSurveyTree}
\title{Generate Survey Tree}
\usage{
generateSurveyTree(shiny = FALSE, blockMap = blockMapParsed)
}
\value{
survey tree
}
\description{
This generates a Survey Tree for a studies WearIT data
}
