% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{getStudyData}
\alias{getStudyData}
\title{getStudyData}
\usage{
getStudyData(
  study_ID = "1045",
  backup_key_file = "~/.auth/.wearit",
  shiny = FALSE,
  apiToken = "",
  base_URL = "https://wearables.vmhost.psu.edu/wearables-survey/api",
  ...
)
}
\arguments{
\item{[String]}{baseURL Base URL, in case you have a variant server}

\item{save}{Boolean indicating whether to save the data to data folder in codebook rmd}
}
\value{
A list containing the auth string and data URL.
}
\description{
Pulls down all the study JSON in raw format.  Mostly, this function just
makes repeated requests and concatenates the JSON data
}
