% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{saveData}
\alias{saveData}
\title{saveData}
\usage{
saveData(data, shiny = FALSE, ...)
}
\arguments{
\item{shiny}{Boolean indicating whether the user is in the shiny app or not}
}
\value{
Saves data to data folder in codebook rmd
}
\description{
This is a wrapper on getStudyData that saves it to the data folder
}
