% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHelpers.R
\name{saveData}
\alias{saveData}
\title{saveData}
\usage{
saveData(pull = FALSE, ...)
}
\arguments{
\item{pull}{whether to pull data from wearIt before saving}

\item{data}{data object to save from}

\item{skip_readline}{Boolean indicating whether the user is in the shiny app and wants to skip readline for api key}
}
\value{
Saves data to data folder in codebook rmd
}
\description{
This is a wrapper on getStudyData that saves it to the data folder
}
